requirements.txt
asgiref==3.9.1
bcrypt==4.1.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==42.0.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0
django-cors-headers==4.5.0
django-debug-toolbar==4.3.0
django-environ==0.12.0
django-pwa==2.0.1
django-templated-mail==1.1.1
django-tenants==3.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.0
gunicorn==21.2.0
idna==3.10
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.6.0
